// src/api/schema.ts
export interface components {
    schemas: {
// ...
        ItemRead: {
            /**
             * Name
             * @example Apple
             */
            name: string;
            /**
             * Note
             * @example Green apple from Japan
             */
            note?: string | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Created By
             * Format: uuid
             */
            created_by: string;
            /** Updated At */
            updated_at?: string | null;
            /** Updated By */
            updated_by?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}